# openapi.yaml
openapi: 3.0.3
info:
  title: Nostr Services API
  description: REST API for Nostr event storage and retrieval with OpenGraph support
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://nostr-rs-api.v0l.io
    description: Production server
paths:
...
  /opengraph/{id}:
    post:
      summary: Inject OpenGraph tags
      description: Inject OpenGraph meta tags into HTML for Nostr events and profiles. This endpoint fetches events and profiles on demand from configured relays to generate rich social media previews.
      operationId: injectOpenGraph
      parameters:
        - name: id
          in: path
          required: true
          description: NIP-19 encoded identifier (note1..., nevent1..., naddr1..., npub1..., nprofile1...)
          schema:
            type: string
          example: "note1..."
        - name: canonical
          in: query
          required: false
          description: URL template for canonical link. Use %s as placeholder for the bech32 ID. For addressable events, uses naddr format.
          schema:
            type: string
          example: "https://example.com/%s"
      requestBody:
        required: true
        description: HTML document to inject meta tags into (max 64 MiB)
        content:
          text/html:
            schema:
              type: string
            example: |
              <!DOCTYPE html>
              <html>
              <head>
              </head>
              <body>
              <div id="root"></div>
              </body>
              </html>
      responses:
        '200':
          description: HTML with injected OpenGraph tags (or unchanged if event not found)
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Request body too large, malformed, or failed to fetch event
...